Moved blog code to a seperate file
